Allow program to read dependencies from egg files
This lets `egg2nix' run with a local egg file or directory, in which
case it will read dependencies from the local egg specification. This
brings it more in line with `cabal2nix' and similar tools, in that
`egg2nix .` will generate a Nix expression from a local source tree.

Adds a test script with some basic checks around input file handling.
The `locate-egg-file' and `read-dependencies' procedures are also
exported for use in the tests, and the `main' entrypoint procedure is
only called when egg2nix.scm is compiled so that it can also be loaded
into the tests or a REPL.